requirements.txt
scipy==1.2.1
numpy==1.14.5
xlwt==1.3.0
requests==2.18.4
lxml==4.2.4
Pillow==6.0.0
beautifulsoup4==4.7.1
